Move format_name.go from packer to builder/common
Per review by pearkes, move format_name.go from packer into builder/common.
